requirements.txt
python-fasthtml
requests
firebase-admin
python-dotenv
ipython
markdown